cargoai/requirements.txt
flask==2.3.3
mail-parser==3.15.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
python-docx==0.8.11
Pillow==10.0.1
transformers==4.35.0
torch==2.3.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
nltk==3.8.1
requests==2.32.3
pytesseract==0.3.10